move functions about schedules to its own module
